[LL-6363] add platform app info panel (#3978)
* create an ExternalLink component

* use new ExternalLink in Alert

* Extract app details from AppCard

* Add title to SideDrawer

* Create InfosDrawer component

* Use InformationDrawer in WebPlatformPlayer TopBar

* move live app info drawer to TopBar

in order to behave the same way as other informations or help drawers

* use default theme values for margin and padding

* use useCallback hook for onClick props

* fix SideDrawer when no title or close button displayed
