+++ src/renderer/actions/UI.js
@@ -6,3 +6,8 @@ export const setTabInformationCenter = createAction("INFORMATION_CENTER_SET_TAB"
   tabId,
 }));
 export const closeInformationCenter = createAction("INFORMATION_CENTER_CLOSE");
+
+export const openPlatformAppInfo = createAction("PLATFORM_APP_INFO_OPEN", manifest => ({
+  manifest,
+}));
+export const closePlatformAppInfo = createAction("PLATFORM_APP_INFO_CLOSE");
