+++ src/renderer/reducers/UI.js
@@ -1,12 +1,19 @@
 // @flow
 
 import { handleActions } from "redux-actions";
+
+import type { AppManifest } from "@ledgerhq/live-common/lib/platform/types";
+
 import type { State } from "~/renderer/reducers";
 
 export type UIState = {
   informationCenter: {
     isOpen: boolean,
     tabId: string,
+  },
+  platformAppInfo: {
+    isOpen: boolean,
+    manifest: ?AppManifest,
   },
 };
 
@@ -15,10 +22,18 @@ const initialState: UIState = {
     isOpen: false,
     tabId: "announcement",
   },
+  platformAppInfo: {
+    isOpen: false,
+    manifest: undefined,
+  },
 };
 
 type OpenPayload = {
   tabId?: string,
+};
+
+type OpenPlatformAppInfoPayload = {
+  manifest: AppManifest,
 };
 
 const handlers = {
@@ -54,6 +69,28 @@ const handlers = {
       },
     };
   },
+
+  PLATFORM_APP_INFO_OPEN: (state, { payload }: { payload: OpenPlatformAppInfoPayload }) => {
+    const { manifest } = payload;
+
+    return {
+      ...state,
+      platformAppInfo: {
+        isOpen: true,
+        manifest,
+      },
+    };
+  },
+
+  PLATFORM_APP_INFO_CLOSE: state => {
+    return {
+      ...state,
+      platformAppInfo: {
+        ...state.platformAppInfo,
+        isOpen: false,
+      },
+    };
+  },
 };
 
 // Selectors
@@ -61,6 +98,8 @@ const handlers = {
 export const UIStateSelector = (state: State): UIState => state.UI;
 
 export const informationCenterStateSelector = (state: Object) => state.UI.informationCenter;
+
+export const platformAppInfoStateSelector = (state: Object) => state.UI.platformAppInfo;
 // Exporting reducer
 
 export default handleActions(handlers, initialState);
